requirements.txt
Flask==2.2.2
numpy==1.21.6
numpy==1.21.5
numpy==1.24.3
opencv_python==4.8.1.78
python-dotenv==1.0.0
tensorflow==1.14.0
tensorflow==2.1.0
tensorflow==2.10.0
tensorflow==2.6.0
tensorflow==2.14.0
tensorflow_intel==2.14.0
